+++ packages/cli/src/main.ts
@@ -2,11 +2,11 @@ import type { Module } from '@snakeskin/language';
 import { createSnakeskinServices, SnakeskinLanguageMetaData } from '@snakeskin/language';
 import chalk from 'chalk';
 import { Command } from 'commander';
-import { extractAstNode } from './util.js';
+import { extractAstNode, extractDocuments } from './util.js';
 import { generateJavaScript } from './generator.js';
 import { NodeFileSystem } from 'langium/node';
 import { startLanguageServer } from 'langium/lsp';
-import { createConnection, ProposedFeatures } from 'vscode-languageserver/node.js';
+import { createConnection, ProposedFeatures, DiagnosticSeverity } from 'vscode-languageserver/node.js';
 import * as url from 'node:url';
 import * as fs from 'node:fs/promises';
 import * as path from 'node:path';
@@ -32,6 +32,43 @@ export const lspAction = () => {
     startLanguageServer(shared);
 }
 
+type ValidateOptions = { Werror: boolean };
+export const validateAction = async (fileOrDirName: string, options: ValidateOptions) => {
+    const services = createSnakeskinServices(NodeFileSystem).Snakeskin;
+    const docs = await extractDocuments(fileOrDirName, services);
+
+    let errorsTotal = 0;
+    let warningsTotal = 0;
+
+    for (const {diagnostics = [], uri} of docs) {
+        const filePath = path.relative(fileOrDirName, uri.path) || fileOrDirName;
+        const errors = diagnostics.filter(diag => diag.severity === DiagnosticSeverity.Error);
+        const warnings = diagnostics.filter(diag => diag.severity === DiagnosticSeverity.Warning);
+
+        if (warnings.length > 0) {
+            warningsTotal += warnings.length;
+            for (const warning of warnings) {
+                console.warn(chalk.yellow(`${filePath}: line ${warning.range.start.line + 1}: ${warning.message}`));
+            }
+        }
+
+        if (errors.length > 0) {
+            errorsTotal += errors.length;
+            for (const validationError of errors) {
+                console.error(chalk.red(`${filePath}: line ${validationError.range.start.line + 1}: ${validationError.message}`));
+            }
+        }
+    }
+
+    const color = errorsTotal > 0 ? chalk.red : warningsTotal > 0 ? chalk.yellow : chalk.green;
+
+    console.log(color(`Found a total of ${errorsTotal} error(s) and ${warningsTotal} warning(s).`))
+
+    if (errorsTotal > 0 || (options.Werror && warningsTotal > 0)) {
+        process.exit(1);
+    }
+}
+
 export default function(): void {
     const program = new Command();
 
@@ -50,6 +87,13 @@ export default function(): void {
         .option('--stdio', 'run in STDIO mode (for compatibility with non-Node clients)')
         .description('runs the standalone LSP server')
         .action(lspAction);
+
+    program
+        .command('validate')
+        .argument('[path]', 'source file or directory; globs not supported yet', '.')
+        .option('-Werror', 'fail on warnings as well', false)
+        .description('parses and validates Snakeskin files')
+        .action(validateAction);
 
     program.parse(process.argv);
 }