+++ packages/extension/CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Some diagnostic messages related to parsing errors around indentation were not being shown.
+- Parser now allows commas to appear at the beginning of plain text
 
 ## [0.0.4] - 2024-08-15