+++ packages/extension/CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.0.5] - 2024-09-16
 
 ### Added
 
@@ -44,7 +44,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 Initial release with basic semantic highlighting and hover information
 
-[Unreleased]: https://github.com/SnakeskinTpl/vscode-snakeskin/compare/v0.0.4...HEAD
+[Unreleased]: https://github.com/SnakeskinTpl/vscode-snakeskin/compare/v0.0.5...HEAD
+[0.0.5]: https://github.com/SnakeskinTpl/vscode-snakeskin/compare/v0.0.4...v0.0.5
 [0.0.4]: https://github.com/SnakeskinTpl/vscode-snakeskin/compare/v0.0.3...v0.0.4
 [0.0.3]: https://github.com/SnakeskinTpl/vscode-snakeskin/compare/v0.0.2...v0.0.3
 [0.0.2]: https://github.com/SnakeskinTpl/vscode-snakeskin/releases/tag/v0.0.2